PreProcessorService/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
kafka-python==2.2.15
MarkupSafe==3.0.2
minio==7.2.15
miniopy-async==1.23.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.2.0
opencv-python==4.10.0.84
openslide-bin==4.0.0.6
openslide-python==1.4.1
pillow==11.0.0
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.1
starlette==0.35.1
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
yarl==1.20.1
python-dotenv==1.0.1


aiokafka==0.12.0
async-timeout==5.0.1
packaging==24.2
tifffile==2025.2.18
tiffslide==2.4.0
zarr==2.18.4

numcodecs==0.15.1
asciitree==0.3.3
Deprecated==1.2.18
wrapt==1.17.2



crc32c==2.7.1
donfig==0.8.1.post1
fasteners==0.19
imagecodecs==2024.12.30

PyYAML==6.0.2

